Refactoring and introduce cmd/convert.go
