+++ cmd/verify.go
